style: component명 figma와 통일
