+++ src/components/mypage/Article/BuyTabArticle/index.tsx
@@ -1,23 +1,21 @@
 import styled from '@emotion/styled'
-import type { SelectOnChangeHandler } from '@offer-ui/react'
-import { Image, SelectBox, Icon, Text, Button } from '@offer-ui/react'
+import { Image, Text, Button } from '@offer-ui/react'
 import type { ReactElement } from 'react'
-import { TRADE_STATUS } from '@constants'
-import type { ArticlesElement, TradeStatus } from '@types'
+import type { TradeBuyActivityType } from '@constants'
+import type { ArticlesElement } from '@types'
 import { toLocaleCurrency } from '@utils'
 
-// NOTE: ArticlesElement 타입에 없어서 임시로 추가한 부분 : likeCount, review, sellerName
-export interface SellTypeArticleProps extends ArticlesElement {
-  // 내 사용자 프로필, 타 사용자 프로필 구분
-  hasToken: boolean
+export interface BuyTabArticleProps extends ArticlesElement {
+  // 구매 활동 타입
+  activityType: TradeBuyActivityType
   className?: string
-  // tradeStatus 변경 시, 이벤트
-  onChangeTradeStatus(productId: number, status: TradeStatus): void
 }
 
-export const SellTypeArticle = (props: SellTypeArticleProps): ReactElement => {
+export const BuyTabArticle = (props: BuyTabArticleProps): ReactElement => {
   const {
+    activityType,
     className,
+    sellerNickName,
     id,
     mainImageUrl,
     title,
@@ -25,47 +23,29 @@ export const SellTypeArticle = (props: SellTypeArticleProps): ReactElement => {
     tradeStatus,
     modifiedDate,
     likeCount,
-    hasToken,
-    isReviewed,
-    onChangeTradeStatus
+    isReviewed
   } = props
-  const isSoldOut = tradeStatus.code === 8
-
-  const handleChangeTradeStatus: SelectOnChangeHandler<TradeStatus> = item => {
-    onChangeTradeStatus?.(id, item)
-  }
+  const isOfferType = activityType === 'offer'
 
   return (
     <StyledContainer className={className}>
-      <StyledProductWrapper hasToken={hasToken}>
+      <StyledProductWrapper>
         <StyledProductImg alt={`product${id}-img`} src={mainImageUrl} />
-        {hasToken ? (
-          <StyledSelectBox
-            items={TRADE_STATUS}
-            value={tradeStatus.code}
-            onChange={handleChangeTradeStatus}
-          />
-        ) : (
-          <StyledFavoriteWrapper isOnlyOther>
-            <Icon color="grayScale50" size={14} type="heart" />
-            <Text styleType="body02R">{likeCount}</Text>
-          </StyledFavoriteWrapper>
-        )}
         <StyledProductMetaWrapper>
+          <StyledSellerName styleType="body02R">
+            {sellerNickName}
+          </StyledSellerName>
           <StyledProductName styleType="body02M">{title}</StyledProductName>
           <StyledProductInfoWrapper>
             <StyledPrice>시작가: {toLocaleCurrency(price)}원</StyledPrice>
-            <StyledFavoriteWrapper isOnlyOther={false}>
-              <Icon color="grayScale50" size={14} type="heart" />
-              <Text styleType="body02R">{likeCount}</Text>
-            </StyledFavoriteWrapper>
-            <StyledDate hasToken={hasToken} styleType="body02R">
-              {modifiedDate}
-            </StyledDate>
+            <StyledTradeStatusName styleType="body02R">
+              {tradeStatus.name}
+            </StyledTradeStatusName>
+            <StyledDate styleType="body02R">{modifiedDate}</StyledDate>
           </StyledProductInfoWrapper>
         </StyledProductMetaWrapper>
       </StyledProductWrapper>
-      {hasToken && isSoldOut && (
+      {isOfferType ? (
         <StyledReviewButtonWrapper>
           <StyledReviewButton
             isReviewed={isReviewed}
@@ -74,6 +54,10 @@ export const SellTypeArticle = (props: SellTypeArticleProps): ReactElement => {
             {isReviewed ? '보낸 후기 보기' : '후기 보내기'}
           </StyledReviewButton>
         </StyledReviewButtonWrapper>
+      ) : (
+        <StyledLikeButton color="grayScale90" size="small" styleType="outline">
+          관심 {likeCount}
+        </StyledLikeButton>
       )}
     </StyledContainer>
   )
@@ -92,17 +76,17 @@ const StyledContainer = styled.li`
     }
   `}
 `
-const StyledProductWrapper = styled.div<{ hasToken: boolean }>`
-  ${({ theme, hasToken }): string => `
+const StyledProductWrapper = styled.div`
+  ${({ theme }): string => `
     display: grid;
     flex: 1;
-    grid-template-columns: ${hasToken ? '90px 90px 1fr' : '90px 1fr'};
+    grid-template-columns: 90px 1fr;
     align-items: center;
     gap: 16px;
     padding: 20px 0 20px 20px;
 
     ${theme.mediaQuery.tablet} {
-      grid-template-columns: 68px 1fr 90px;
+      grid-template-columns: 68px 1fr;
       gap: 8px;
       padding: 16px 24px;
     }
@@ -117,7 +101,6 @@ const StyledProductImg = styled(Image)`
   ${({ theme }): string => `
     width: 90px;
     height: 90px;
-    order:1;
 
     ${theme.mediaQuery.tablet} {
       width: 68px;
@@ -125,14 +108,20 @@ const StyledProductImg = styled(Image)`
     }
   `}
 `
-const StyledSelectBox = styled(SelectBox)`
+const StyledSellerName = styled(Text)`
   ${({ theme }): string => `
-    order: 2;
+    text-align: center;
+    color: ${theme.colors.grayScale70};
+    max-width: 100px;
+    text-overflow: ellipsis;
+    overflow: hidden;
+    white-space: nowrap;
+    word-break: break-word;
 
-    ${theme.mediaQuery.tablet} {
-      order: 3;
-    }
-  `}
+  ${theme.mediaQuery.tablet} {
+    ${theme.fonts.caption01M};
+  }
+`}
 `
 
 const StyledProductMetaWrapper = styled.div`
@@ -140,13 +129,10 @@ const StyledProductMetaWrapper = styled.div`
     display: flex;
     align-items: center;
     justify-content: space-around;
-    order: 3;
 
     ${theme.mediaQuery.tablet} {
       align-items: flex-start;
       flex-direction: column;
-      order: 2;
-      gap: 4px;
     }
   `}
 `
@@ -161,11 +147,12 @@ const StyledProductName = styled(Text)`
 
     ${theme.mediaQuery.tablet} {
       text-align: left;
-      width: 460px;
+      max-width: 460px;
+      margin-bottom: 6px;
     }
 
     ${theme.mediaQuery.mobile} {
-      width: 171px;
+      max-width: 171px;
     }
   `}
 `
@@ -193,28 +180,26 @@ const StyledPrice = styled.span`
     }
   `}
 `
-const StyledFavoriteWrapper = styled.div<{ isOnlyOther: boolean }>`
-  ${({ theme, isOnlyOther }): string => `
-    display: ${isOnlyOther ? 'none' : 'flex'};
+const StyledTradeStatusName = styled(Text)`
+  ${({ theme }): string => `
+    display: flex;
+    flex-direction: column;
     align-items: center;
-    justify-content: center;
-    gap: 2px;
     width: 100px;
     color: ${theme.colors.grayScale50};
 
     ${theme.mediaQuery.tablet} {
-      display: ${isOnlyOther ? 'flex' : 'none'};
-      order: 3;
+      display: none;
     }
   `}
 `
-const StyledDate = styled(Text)<{ hasToken: boolean }>`
-  ${({ theme, hasToken }): string => `
+const StyledDate = styled(Text)`
+  ${({ theme }): string => `
     display: inline-block;
     color: ${theme.colors.grayScale50};
 
     ${theme.mediaQuery.tablet} {
-      display: ${hasToken ? 'none' : 'inline-block'};
+      display: none;
       ${theme.fonts.caption01M};
       color: ${theme.colors.grayScale30};
     }
@@ -245,6 +230,16 @@ const StyledReviewButton = styled(Button)<{ isReviewed: boolean }>`
       border-radius: 0;
       padding: 20px 0;
       margin-right: 0;
+    }
+  `}
+`
+const StyledLikeButton = styled(Button)`
+  ${({ theme }): string => `
+    color: ${theme.colors.grayScale90};
+    margin-right: 20px;
+
+    ${theme.mediaQuery.tablet} {
+      display: none;
     }
   `}
 `
