+++ src/components/mypage/Article/SellTabArticle/index.tsx
@@ -1,21 +1,23 @@
 import styled from '@emotion/styled'
-import { Image, Text, Button } from '@offer-ui/react'
+import type { SelectOnChangeHandler } from '@offer-ui/react'
+import { Image, SelectBox, Icon, Text, Button } from '@offer-ui/react'
 import type { ReactElement } from 'react'
-import type { TradeBuyActivityType } from '@constants'
-import type { ArticlesElement } from '@types'
+import { TRADE_STATUS } from '@constants'
+import type { ArticlesElement, TradeStatus } from '@types'
 import { toLocaleCurrency } from '@utils'
 
-export interface BuyTypeArticleProps extends ArticlesElement {
-  // 구매 활동 타입
-  activityType: TradeBuyActivityType
+// NOTE: ArticlesElement 타입에 없어서 임시로 추가한 부분 : likeCount, review, sellerName
+export interface SellTabArticleProps extends ArticlesElement {
+  // 내 사용자 프로필, 타 사용자 프로필 구분
+  hasToken: boolean
   className?: string
+  // tradeStatus 변경 시, 이벤트
+  onChangeTradeStatus(productId: number, status: TradeStatus): void
 }
 
-export const BuyTypeArticle = (props: BuyTypeArticleProps): ReactElement => {
+export const SellTabArticle = (props: SellTabArticleProps): ReactElement => {
   const {
-    activityType,
     className,
-    sellerNickName,
     id,
     mainImageUrl,
     title,
@@ -23,29 +25,47 @@ export const BuyTypeArticle = (props: BuyTypeArticleProps): ReactElement => {
     tradeStatus,
     modifiedDate,
     likeCount,
-    isReviewed
+    hasToken,
+    isReviewed,
+    onChangeTradeStatus
   } = props
-  const isOfferType = activityType === 'offer'
+  const isSoldOut = tradeStatus.code === 8
+
+  const handleChangeTradeStatus: SelectOnChangeHandler<TradeStatus> = item => {
+    onChangeTradeStatus?.(id, item)
+  }
 
   return (
     <StyledContainer className={className}>
-      <StyledProductWrapper>
+      <StyledProductWrapper hasToken={hasToken}>
         <StyledProductImg alt={`product${id}-img`} src={mainImageUrl} />
+        {hasToken ? (
+          <StyledSelectBox
+            items={TRADE_STATUS}
+            value={tradeStatus.code}
+            onChange={handleChangeTradeStatus}
+          />
+        ) : (
+          <StyledFavoriteWrapper isOnlyOther>
+            <Icon color="grayScale50" size={14} type="heart" />
+            <Text styleType="body02R">{likeCount}</Text>
+          </StyledFavoriteWrapper>
+        )}
         <StyledProductMetaWrapper>
-          <StyledSellerName styleType="body02R">
-            {sellerNickName}
-          </StyledSellerName>
           <StyledProductName styleType="body02M">{title}</StyledProductName>
           <StyledProductInfoWrapper>
             <StyledPrice>시작가: {toLocaleCurrency(price)}원</StyledPrice>
-            <StyledTradeStatusName styleType="body02R">
-              {tradeStatus.name}
-            </StyledTradeStatusName>
-            <StyledDate styleType="body02R">{modifiedDate}</StyledDate>
+            <StyledFavoriteWrapper isOnlyOther={false}>
+              <Icon color="grayScale50" size={14} type="heart" />
+              <Text styleType="body02R">{likeCount}</Text>
+            </StyledFavoriteWrapper>
+            <StyledDate hasToken={hasToken} styleType="body02R">
+              {modifiedDate}
+            </StyledDate>
           </StyledProductInfoWrapper>
         </StyledProductMetaWrapper>
       </StyledProductWrapper>
-      {isOfferType ? (
+      {hasToken && isSoldOut && (
         <StyledReviewButtonWrapper>
           <StyledReviewButton
             isReviewed={isReviewed}
@@ -54,10 +74,6 @@ export const BuyTypeArticle = (props: BuyTypeArticleProps): ReactElement => {
             {isReviewed ? '보낸 후기 보기' : '후기 보내기'}
           </StyledReviewButton>
         </StyledReviewButtonWrapper>
-      ) : (
-        <StyledLikeButton color="grayScale90" size="small" styleType="outline">
-          관심 {likeCount}
-        </StyledLikeButton>
       )}
     </StyledContainer>
   )
@@ -76,17 +92,17 @@ const StyledContainer = styled.li`
     }
   `}
 `
-const StyledProductWrapper = styled.div`
-  ${({ theme }): string => `
+const StyledProductWrapper = styled.div<{ hasToken: boolean }>`
+  ${({ theme, hasToken }): string => `
     display: grid;
     flex: 1;
-    grid-template-columns: 90px 1fr;
+    grid-template-columns: ${hasToken ? '90px 90px 1fr' : '90px 1fr'};
     align-items: center;
     gap: 16px;
     padding: 20px 0 20px 20px;
 
     ${theme.mediaQuery.tablet} {
-      grid-template-columns: 68px 1fr;
+      grid-template-columns: 68px 1fr 90px;
       gap: 8px;
       padding: 16px 24px;
     }
@@ -101,6 +117,7 @@ const StyledProductImg = styled(Image)`
   ${({ theme }): string => `
     width: 90px;
     height: 90px;
+    order:1;
 
     ${theme.mediaQuery.tablet} {
       width: 68px;
@@ -108,20 +125,14 @@ const StyledProductImg = styled(Image)`
     }
   `}
 `
-const StyledSellerName = styled(Text)`
+const StyledSelectBox = styled(SelectBox)`
   ${({ theme }): string => `
-    text-align: center;
-    color: ${theme.colors.grayScale70};
-    max-width: 100px;
-    text-overflow: ellipsis;
-    overflow: hidden;
-    white-space: nowrap;
-    word-break: break-word;
+    order: 2;
 
-  ${theme.mediaQuery.tablet} {
-    ${theme.fonts.caption01M};
-  }
-`}
+    ${theme.mediaQuery.tablet} {
+      order: 3;
+    }
+  `}
 `
 
 const StyledProductMetaWrapper = styled.div`
@@ -129,10 +140,13 @@ const StyledProductMetaWrapper = styled.div`
     display: flex;
     align-items: center;
     justify-content: space-around;
+    order: 3;
 
     ${theme.mediaQuery.tablet} {
       align-items: flex-start;
       flex-direction: column;
+      order: 2;
+      gap: 4px;
     }
   `}
 `
@@ -147,12 +161,11 @@ const StyledProductName = styled(Text)`
 
     ${theme.mediaQuery.tablet} {
       text-align: left;
-      max-width: 460px;
-      margin-bottom: 6px;
+      width: 460px;
     }
 
     ${theme.mediaQuery.mobile} {
-      max-width: 171px;
+      width: 171px;
     }
   `}
 `
@@ -180,26 +193,28 @@ const StyledPrice = styled.span`
     }
   `}
 `
-const StyledTradeStatusName = styled(Text)`
-  ${({ theme }): string => `
-    display: flex;
-    flex-direction: column;
+const StyledFavoriteWrapper = styled.div<{ isOnlyOther: boolean }>`
+  ${({ theme, isOnlyOther }): string => `
+    display: ${isOnlyOther ? 'none' : 'flex'};
     align-items: center;
+    justify-content: center;
+    gap: 2px;
     width: 100px;
     color: ${theme.colors.grayScale50};
 
     ${theme.mediaQuery.tablet} {
-      display: none;
+      display: ${isOnlyOther ? 'flex' : 'none'};
+      order: 3;
     }
   `}
 `
-const StyledDate = styled(Text)`
-  ${({ theme }): string => `
+const StyledDate = styled(Text)<{ hasToken: boolean }>`
+  ${({ theme, hasToken }): string => `
     display: inline-block;
     color: ${theme.colors.grayScale50};
 
     ${theme.mediaQuery.tablet} {
-      display: none;
+      display: ${hasToken ? 'none' : 'inline-block'};
       ${theme.fonts.caption01M};
       color: ${theme.colors.grayScale30};
     }
@@ -230,16 +245,6 @@ const StyledReviewButton = styled(Button)<{ isReviewed: boolean }>`
       border-radius: 0;
       padding: 20px 0;
       margin-right: 0;
-    }
-  `}
-`
-const StyledLikeButton = styled(Button)`
-  ${({ theme }): string => `
-    color: ${theme.colors.grayScale90};
-    margin-right: 20px;
-
-    ${theme.mediaQuery.tablet} {
-      display: none;
     }
   `}
 `
