+++ src/components/mypage/ArticleList/BuyTabArticleList/index.tsx
@@ -1,27 +1,27 @@
 import { Divider } from '@offer-ui/react'
 import type { ReactElement } from 'react'
 import { Fragment } from 'react'
-import { BuyTypeArticle } from '../../Article'
+import { BuyTabArticle } from '../../Article'
 import type { TradeBuyActivityType } from '@constants'
 import type { ArticlesElement } from '@types'
 
-export interface BuyTypeArticleListProps {
+export interface BuyTabArticleListProps {
   activityType: TradeBuyActivityType
   articles: ArticlesElement[]
   className?: string
 }
 
-export const BuyTypeArticleList = ({
+export const BuyTabArticleList = ({
   activityType,
   articles,
   className
-}: BuyTypeArticleListProps): ReactElement => {
+}: BuyTabArticleListProps): ReactElement => {
   return (
     <ul className={className}>
       {articles.map((article, index) => {
         return (
           <Fragment key={article.id}>
-            <BuyTypeArticle activityType={activityType} {...article} />
+            <BuyTabArticle activityType={activityType} {...article} />
             {index !== articles.length - 1 && <Divider size="bold" />}
           </Fragment>
         )
