+++ src/components/mypage/ArticleList/ReviewTabArticleList/index.tsx
@@ -1,23 +1,23 @@
 import { Divider } from '@offer-ui/react'
 import type { ReactElement } from 'react'
 import { Fragment } from 'react'
-import { ReviewTypeArticle } from '@components/mypage/Article'
+import { ReviewTabArticle } from '@components/mypage/Article'
 import type { ReviewsElement } from '@types'
 
-export interface ReviewTypeArticleListProps {
+export interface ReviewTabArticleListProps {
   reviews: ReviewsElement[]
   className?: string
 }
 
-export const ReviewTypeArticleList = ({
+export const ReviewTabArticleList = ({
   reviews,
   className
-}: ReviewTypeArticleListProps): ReactElement => {
+}: ReviewTabArticleListProps): ReactElement => {
   return (
     <ul className={className}>
       {reviews.map((review, index) => (
         <Fragment key={review.id}>
-          <ReviewTypeArticle {...review} />
+          <ReviewTabArticle {...review} />
           {index !== reviews.length - 1 && <Divider />}
         </Fragment>
       ))}
